Add some text if theres no data about selected mag
